Split library & file list handlers for readability
